refactor wizard page theme
